projections: Refactor rabbit_khepri_bindings table layout
The old shape used the source exchange as the key in a bag table. This
made it cheaper to list all bindings for an exchange but it caused a
bad regression for inserting many bindings, for example in a definition
import. If all bindings are for the same exchange, inserting N bindings
has exponential time complexity when using a bag compared to the
roughly linear time complexity using a set.

This reflects the shape of the `rabbit_route` table from mnesia: the
whole `#binding{}` record is the key of the table and the table is
a set. With this change, the time taken to import definitions roughly
matches mnesia.