Khepri: Limit the index route projection to direct/fanout bindings
The parent commit introduced the type of the exchange as a queryable
piece of data on the exchange component in the rabbit_db_binding
subtree. We can use that information to limit the bindings in the
'rabbit_khepri_index_route' projection to only direct or fanout
bindings. This might reduce memory usage of that table when using
many non-direct and non-fanout bindings.