Khepri: Use queue projection for rabbit_db_queue:get/1
This change uses the Khepri projection for queues for
rabbit_db_queue:get/1 and rabbit_db_queue:get_durable/1.
Switching to the projection here for 'get's improves performance for
any operation that needs to look up a queue. For example, importing
a set of 100k classic queues takes a noticeably smaller amount of time.

| Q getting function  | Time to import 100k classic queues (secs) |
|---                  |---                                        |
| rabbit_khepri:get/1 | 19.20                                     |
| ets:lookup/2        | 15.71                                     |
| (mnesia)            | 14.96                                     |